# gateway/docs/swagger.yaml
definitions:
  genproto.AddBookingReq:
    properties:
      location:
        $ref: '#/definitions/genproto.GeoPoint'
      provider_id:
        type: string
      schudule_time:
        type: string
      service_id:
        type: string
      user_id:
        type: string
    type: object
  genproto.AddNotificationReq:
    properties:
      message:
        type: string
      user_id:
        type: string
    type: object
...
  genproto.Booking:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      location:
        $ref: '#/definitions/genproto.GeoPoint'
      provider_id:
        $ref: '#/definitions/genproto.Provider'
      schudule_time:
        type: string
      service_id:
        $ref: '#/definitions/genproto.Services'
      status:
        type: string
      total_price:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  genproto.BookingsByProviderResp:
    properties:
      bookings:
        items:
          $ref: '#/definitions/genproto.Booking'
        type: array
    type: object
  genproto.Cart:
    properties:
      id:
        type: string
      number:
        type: integer
      service_id:
        $ref: '#/definitions/genproto.Services'
      total:
        type: integer
      user_id:
        type: string
    type: object
  genproto.CreateCartReq:
    properties:
      number:
        type: integer
      service_id:
        type: string
      total:
        type: integer
      user_id:
        type: string
    type: object
  genproto.GeoPoint:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
  genproto.GetBookingResp:
    properties:
      booking:
        $ref: '#/definitions/genproto.Booking'
    type: object
  genproto.GetNotificationsResp:
    properties:
      notifications:
        items:
          $ref: '#/definitions/genproto.Notification'
        type: array
    type: object
...
  genproto.ListAllBookingsResp:
    properties:
      bookings:
        items:
          $ref: '#/definitions/genproto.Booking'
        type: array
    type: object
...
  genproto.MarkNotificationAsReadReq:
    properties:
      id:
        type: string
    type: object
  genproto.Notification:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_read:
        type: boolean
      message:
        type: string
      user_id:
        type: string
    type: object
...
  genproto.Provider:
    properties:
      availability:
        type: string
      average_rating:
        type: number
      company_name:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: string
      location:
        $ref: '#/definitions/genproto.GeoPoint'
      services:
        items:
          type: string
        type: array
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  genproto.Services:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      duration:
        type: integer
      id:
        type: string
      name:
        type: string
      price:
        type: number
      updated_at:
        type: string
    type: object
  genproto.UpdateBookingReq:
    properties:
      booking:
        $ref: '#/definitions/genproto.Booking'
      id:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server.
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /api/v1/booking:
    get:
      consumes:
      - application/json
      description: This API lists all bookings with optional filters
      parameters:
      - description: User ID
        in: query
        name: user_id
        type: string
      - description: Status
        in: query
        name: status
        type: string
      - description: Provider ID
        in: query
        name: provider_id
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.ListAllBookingsResp'
        "400":
          description: 'error": "error description'
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List all bookings
      tags:
      - carwash/booking
  /api/v1/booking/{id}:
    delete:
      consumes:
      - application/json
      description: This API deletes a booking by its ID
      parameters:
      - description: Booking Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message": "booking deleted successfully'
          schema:
            type: string
        "400":
          description: 'error": "error description'
          schema:
            type: string
        "500":
          description: 'error": "error description'
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete booking
      tags:
      - carwash/booking
    get:
      consumes:
      - application/json
      description: This API retrieves a booking by its ID
      parameters:
      - description: Booking Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.GetBookingResp'
        "400":
          description: 'error": "error description'
          schema:
            type: string
        "500":
          description: 'error": "error description'
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get booking
      tags:
      - carwash/booking
    put:
      consumes:
      - application/json
      description: This API updates an existing booking
      parameters:
      - description: Booking Id
        in: path
        name: id
        required: true
        type: string
      - description: Update booking Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/genproto.UpdateBookingReq'
      produces:
      - application/json
      responses:
        "200":
          description: 'message": "booking updated successfully'
          schema:
            type: string
        "400":
          description: 'error": "error description'
          schema:
            type: string
        "500":
          description: 'error": "error description'
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update booking
      tags:
      - carwash/booking
  /api/v1/booking/add:
    post:
      consumes:
      - application/json
      description: This API adds a new booking
      parameters:
      - description: Add booking Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/genproto.AddBookingReq'
      produces:
      - application/json
      responses:
        "200":
          description: 'message": "booking created successfully'
          schema:
            type: string
        "400":
          description: 'error": "error description'
          schema:
            type: string
        "500":
          description: 'error": "error description'
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Add booking
      tags:
      - carwash/booking
  /api/v1/booking/provider:
    get:
      consumes:
      - application/json
      description: This API retrieves bookings for a specific provider
      parameters:
      - description: Provider ID
        in: query
        name: provider_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.BookingsByProviderResp'
        "400":
          description: 'error": "error description'
          schema:
            type: string
        "500":
          description: 'error": "error description'
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get bookings by provider
      tags:
      - carwash/booking
  /api/v1/cart/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific cart by its ID.
      parameters:
      - description: Cart ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Cart'
        "400":
          description: 'error": "error message'
          schema:
            type: string
      summary: GET cart by ID
      tags:
      - carwash/Cart
  /api/v1/cart/add:
    post:
      consumes:
      - application/json
      description: This API creates a new cart.
      parameters:
      - description: Cart data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/genproto.CreateCartReq'
      produces:
      - application/json
      responses:
        "200":
          description: 'message": "created successfully'
          schema:
            type: string
        "400":
          description: 'error": "error message'
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: CREATE cart
      tags:
      - carwash/Cart
  /api/v1/notification/{id}:
    get:
      consumes:
      - application/json
      description: This API retrieves a notification by ID
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.GetNotificationsResp'
        "400":
          description: 'error": "error message'
          schema:
            type: string
      summary: GET notification
      tags:
      - carwash/Notification
  /api/v1/notification/{id}/read:
    put:
      consumes:
      - application/json
      description: This API marks a notification as read
      parameters:
      - description: Notification
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/genproto.MarkNotificationAsReadReq'
      produces:
      - application/json
      responses:
        "200":
          description: 'message": "updated successfully'
          schema:
            type: string
        "400":
          description: 'error": "error message'
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: MARK notification as read
      tags:
      - carwash/Notification
  /api/v1/notification/add:
    post:
      consumes:
      - application/json
      description: This API creates a notification
      parameters:
      - description: Notification
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/genproto.AddNotificationReq'
      produces:
      - application/json
      responses:
        "200":
          description: 'message": "created successfully'
          schema:
            type: string
        "400":
          description: 'error": "error message'
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: CREATE notification
      tags:
      - carwash/Notification
...
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"